// types/hellofresh.ts
export interface Search {
  take: number;
  skip: number;
  count: number;
  total: number;
  items: Item[];
}

export interface Item {
  id: string;
  country: Country;
  name: string;
  seoName: string;
  category: Category | null;
  slug: string;
  headline: string;
  description: string;
  descriptionHTML: string;
  descriptionMarkdown: string;
  seoDescription: string;
  comment: null | string;
  difficulty: number;
  prepTime: string;
  totalTime: string;
  servingSize: number;
  createdAt: string;
  updatedAt: string;
  link: string;
  imageLink: string;
  imagePath: string;
  cardLink: null;
  videoLink: null;
  nutrition: Nutrition[];
  ingredients: ItemIngredient[];
  allergens: Allergen[];
  utensils: Utensil[];
  tags: Tag[];
  cuisines: Category[];
  yields: Yield[];
  steps: Step[];
  wines: any[];
  marketplaceItems: any[];
  author: Author;
  label: Label | null;
  yieldType: string;
  averageRating: number | null;
  ratingsCount: number;
  favoritesCount: number;
  active: boolean;
  highlighted: boolean;
  isDinnerToLunch: boolean;
  isExcludedFromIndex: boolean;
  isAddon: boolean;
  isPremium: boolean;
  websiteUrl: string;
  clonedFrom: string;
  canonical: null;
  canonicalLink: null;
  promotion: null;
  uniqueRecipeCode: string;
  uuid: string;
}

export interface Allergen {
  id: string;
  type: string;
  description: null;
  tracesOf: boolean;
  triggersTracesOf: boolean;
  iconLink: null | string;
  iconPath: null | string;
  usage: number;
  name: string;
  slug: string;
}

export enum Author {
  Empty = "",
  Nieuw = "nieuw",
}

export interface Category {
  id: string;
  type: string;
  name: string;
  slug: string;
  iconLink: null | string;
  iconPath: string;
  usage: number;
}

export enum Country {
  Fr = "FR",
}

export interface ItemIngredient {
  country: Country;
  id: string;
  type: string;
  name: string;
  slug: string;
  description: null;
  internalName: null;
  shipped: boolean;
  imageLink: null | string;
  imagePath: string;
  usage: number;
  hasDuplicatedName: null;
  allergens: string[];
  family: Family;
  uuid: string;
}

export interface Family {
  id: ID;
  type: FamilyType;
  description: null | string;
  priority: number;
  iconLink: string;
  iconPath: FamilyIconPath;
  usageByCountry: { [key: string]: number };
  createdAt: string;
  updatedAt: string;
  uuid: string;
  name: SlugEnum;
  slug: SlugEnum;
}

export enum FamilyIconPath {
  IngredientFamilies5D481C077045C100115Aa09F753246A9PNG = "/ingredient-families/5d481c077045c100115aa09f-753246a9.png",
  IngredientFamilies6299E0F4F7Cd25Bc1C02438411Bf4Ba9PNG = "/ingredient-families/6299e0f4f7cd25bc1c024384-11bf4ba9.png",
}

export enum ID {
  The5D481C077045C100115Aa09F = "5d481c077045c100115aa09f",
  The6299E0F4F7Cd25Bc1C024384 = "6299e0f4f7cd25bc1c024384",
}

export enum SlugEnum {
  None = "none",
  RepasKit = "repas-kit",
}

export enum FamilyType {
  BreakfastRecipes = "breakfast-recipes",
  None = "none",
}

export interface Label {
  text: string;
  handle: Author;
  backgroundColor: string;
  foregroundColor: string;
  displayLabel: boolean;
}

export interface Nutrition {
  type: NutritionType;
  name: NutritionName;
  amount: number;
  unit: NutritionUnit;
}

export enum NutritionName {
  DontAcidesGrasSaturés = "dont acides gras saturés",
  DontSucres = "dont sucres",
  Glucides = "Glucides",
  MatièresGrasses = "Matières grasses",
  Protéines = "Protéines",
  Sel = "Sel",
  ÉnergieKJ = "Énergie (kJ)",
  ÉnergieKcal = "Énergie (kcal)",
}

export enum NutritionType {
  The57B42A48B7E8697D4B305304 = "57b42a48b7e8697d4b305304",
  The57B42A48B7E8697D4B305305 = "57b42a48b7e8697d4b305305",
  The57B42A48B7E8697D4B305306 = "57b42a48b7e8697d4b305306",
  The57B42A48B7E8697D4B305307 = "57b42a48b7e8697d4b305307",
  The57B42A48B7E8697D4B305308 = "57b42a48b7e8697d4b305308",
  The57B42A48B7E8697D4B305309 = "57b42a48b7e8697d4b305309",
  The57B42A48B7E8697D4B30530B = "57b42a48b7e8697d4b30530b",
  The57B42A48B7E8697D4B30530D = "57b42a48b7e8697d4b30530d",
}

export enum NutritionUnit {
  G = "g",
  KJ = "kJ",
  Kcal = "kcal",
}

export interface Step {
  index: number;
  instructionsMarkdown: string;
  instructionsHTML: string;
  instructions: string;
  timers: any[];
  images: Image[];
  videos: any[];
  ingredients: any[];
  utensils: string[];
}

export interface Image {
  link: string;
  path: string;
  caption: string;
}

export interface Tag {
  id: string;
  type: string;
  iconLink: null | string;
  iconPath: TagIconPath | null;
  numberOfRecipes: number;
  numberOfRecipesByCountry: { [key: string]: number };
  colorHandle: string;
  preferences: Preference[];
  name: string;
  slug: string;
  displayLabel: boolean;
}

export enum TagIconPath {
  Empty = "",
  Tags619F65Dae9Aadf39547F9F657Bd41480PNG = "/tags/619f65dae9aadf39547f9f65-7bd41480.png",
}

export enum Preference {
  CalorieSmart = "Calorie Smart",
  Classique = "Classique",
  Famille = "Famille",
  Flexitarien = "Flexitarien",
  Original = "Original",
  PoissonLégumes = "Poisson & Légumes",
  Rapido = "Rapido",
  Végétarien = "Végétarien",
}

export interface Utensil {
  id: string;
  type: null | string;
  name: string;
}

export interface Yield {
  yields: number;
  ingredients: YieldIngredient[];
}

export interface YieldIngredient {
  id: string;
  amount: number | null;
  unit: IngredientUnit;
}